dns_app/requirements.txt
Flask~=3.0.2
dnspython~=2.6.1
requests~=2.31.0